Add files via upload
In this project two dataset has been analyzed. First the data set of Hospital visit through out existing departments. I am trying to analyze the relationship between the average minutes to service and the average number of days spent in the hospital, broken down by department type and doctor. the second  is the data set of data science colleges and universities, which is available here:https://github.com/ryanswanstrom/awesome-datascience-colleges .